requirements.txt
Flask==2.2.2
opencv-python-headless==4.8.0.74
numpy==1.24.3
Pillow==9.5.0
dlib==19.24.2
